improved `Geometry` abstraction

+++ src/innovus/assets/default3d.v.glsl
@@ -7,7 +7,7 @@ uniform sampler2D tex_atlas;
 
 layout(location = 0) in vec3 vertex_pos;
 layout(location = 1) in vec4 vertex_color;
-layout(location = 2) in float vertex_tex;
+layout(location = 2) in uint vertex_tex;
 layout(location = 3) in vec2 vertex_uv;
 layout(location = 4) in vec3 vertex_norm;
 
@@ -22,7 +22,7 @@ out Vertex {
 void main() {
     vertex.pos = vertex_pos;
     vertex.color = vertex_color;
-    vertex.tex = vertex_tex;
+    vertex.tex = float(vertex_tex);
     vertex.uv = vertex_uv;
     vertex.norm = vertex_norm;
